+++ modules/flann/include/opencv2/flann/nn_index.h
@@ -114,7 +114,7 @@ public:
         int* indices_ptr = NULL;
         DistanceType* dists_ptr = NULL;
         if (indices.cols > 0) {
-            n = indices.cols;
+            n = (int)indices.cols;
             indices_ptr = indices[0];
             dists_ptr = dists[0];
         }
@@ -127,7 +127,7 @@ public:
             else resultSet.copy(indices_ptr, dists_ptr, n);
         }
 
-        return resultSet.size();
+        return (int)resultSet.size();
     }
 
     /**
